redux ducks, reducers/actions made, need to connect components to redux, added log in route to signup, combine reducers, need to refactor and finish connection
