+++ client/src/reducers/index.js
@@ -1,0 +1,10 @@
+import { combineReducers } from 'redux';
+import login from './loginReducer';
+import signUp from './signUpReducer';
+
+
+
+export default combineReducers({
+    login,
+    signUp,
+});
